third_party/WebKit/LayoutTests/imported/web-platform-tests/html/infrastructure/common-dom-interfaces/collections/htmloptionscollection-expected.txt

This is a testharness.js-based test.
PASS Original length 
PASS Setting length to original value has no effect 
PASS Setting length to shorter value 
PASS Setting length to longer value 
FAIL Insert <p><option>6</option></p> into <select> assert_equals: Correct length expected 5 but got 6
FAIL Insert <select><option>6</option></select> into <select> assert_equals: Correct length expected 5 but got 6
PASS Insert <optgroup><option>6</option></optgroup> into <select> 
FAIL Insert <optgroup><optgroup><option>6</option></optgroup></optgroup> into <select> assert_equals: Correct length expected 5 but got 6
PASS namedItem id attribute 
PASS namedItem name attribute 
PASS namedItem doesn't match anything 
FAIL namedItem multiple IDs assert_equals: expected (string) "3" but got (undefined) undefined
FAIL namedItem multiple names assert_equals: expected (string) "5" but got (undefined) undefined
FAIL namedItem multiple name and ID assert_equals: expected (string) "7" but got (undefined) undefined
FAIL namedItem multiple name and ID with multiple attributes assert_equals: expected (string) "9" but got (undefined) undefined
PASS namedItem id attribute multiple attributes one element 
PASS namedItem name attribute multiple attributes one element 
PASS Add non-option to collection 
Harness: the test ran to completion.
